+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `replaceInjectablesOverrides` prop to `OneBlinkFormBase`
+- `description` to `ArcGISWebMapElement` graphic attribute options
 
 ## [8.4.0] - 2025-06-03